==> 0_first_look.ipynb <==
##0. Initial look into crime data

Purpose:
* import data files and merge important features into one dataset

import pandas as pd
import numpy as np
import unicodedata

We want to have the latitude and longitude for all crime events. The info we need is:
* report number
* report date/time
* address
* zip
* lat
* long
* NIBRS code

# read in all the excel files with the first code styling

df = pd.ExcelFile("data/1993 AND 1994.xlsx").parse(parse_dates= True)
df = df.append(pd.ExcelFile("data/1995 AND 1996.xlsx").parse())
df = df.append(pd.ExcelFile("data/1997 AND 1998.xlsx").parse())
df = df.append(pd.ExcelFile("data/1999 AND 2000.xlsx").parse())
df = df.append(pd.ExcelFile("data/2001 AND 2002.xlsx").parse())
df = df.append(pd.ExcelFile("data/2003 AND 2004.xlsx").parse())
df = df.append(pd.ExcelFile("data/2005 THROUGH 2007.xlsx").parse())

# rename the columns, we have more columns than we need
print len(df)
print df.columns.values
df.columns = ['report_number','address','code_text','datetime_occurred','district','nibrs_code','long','lat','zip','nature']
print df.columns.values

# read the second style of crime reporting excel files 

df2 = pd.ExcelFile("data/2007 REPORT DATA.xlsx").parse()
df2 = df2.append(pd.ExcelFile("data/2008 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2009 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2010 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2011 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2012 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2013 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2014 REPORT DATA.xlsx").parse())
df2 = df2.append(pd.ExcelFile("data/2015 REPORT DATA.xlsx").parse())

# rename the columns, we have more columns than we need, and different ones!

print len(df2)
print df2.columns.values
df2.columns = ['report_number','datetime_reported','address','apt','city','zip','district','datetime_occurred','datetime_to','nibrs_code','nature']
print df2.columns.values

# combine the two dataframes, including columns will just lead to the creation of NaN entries

df3 = df[['report_number','address','datetime_occurred','district','nibrs_code','long','lat','zip']].append( \
      df2[['report_number','address','city','datetime_occurred','district','nibrs_code','zip']])

# how many null values are there?

print len(df3)
print df3.isnull().sum()

# capitalize the nibrs_code and the city fields 

df3.loc[~df3['nibrs_code'].isnull(),'nibrs_code'] = df3.loc[~df3['nibrs_code'].isnull(),'nibrs_code'].apply(lambda x: str(x).upper())
df3.loc[~df3['city'].isnull(),'city'] =df3.loc[~df3['city'].isnull(),'city'].apply(lambda x: str(x).upper())

# make the zip field a string
def clean_zip(x):
    try:
        x = str(int(x))
    except ValueError:
        pass
    return x  

df3['zip'] = df3['zip'].apply(clean_zip)

# make the address uppercase as well
def uppercase(s):
    try:
        s = s.upper()
    except AttributeError:
        pass
    except TypeError:
        pass
    return s

df3['address'] = df3['address'].apply(uppercase)

Making these fields uppercase allows us to filter common entries easier!

df3.info()

df3.to_csv('data/summit_crime_ungeocoded.csv',index=False, encoding='utf-8')
