==> 1_filter_by_nibrs_code.ipynb <==
#1. Filter data by report number and NIBRS code

Purpose:
* filter the entire data set by report number and NIBRS code and output 2 datasets
* output a dataset containing the entire dataset filtered, from which we would like to geocode
* output a dataset containing the set of unique addresses from the filtered set, from which we would like to geocode
* data contains multiple NIBRS codes for each report number, however this does not represent a unique event. We have currently decided to filter by the most "severe" NIBRS code, which is given in a printout, but also described by sorting the codes alphabetically (closer to the beginning, starting with 0, is more severe)

import pandas as pd

df = pd.read_csv('data/summit_crime_ungeocoded.csv')

df.info()

# must have a NIBRS code
df = df[~df['nibrs_code'].isnull()]

df.info()

## Filter the NIBRS code for each report number

# the order of the list 'nibrs' is the order of severity of crime, so a 0 takes precedence over a 1, etc. 
nibrs= ['09A','09B','09C','100','11A','11B','11C','11D','120','13A','13B','13C','200','210','220','23A','23B','23C','23D','23E','23F', \
        '23G','23H','240','250','26A','26B','26C','26D','26E','270','280','290','35A','35B','36A','36B','370','39A','39B','39C','39D', \
        '40A','40B','510','520','90A','90B','90C','90D','90E','90F','90G','90H','90I','90J','90Z']
# for easier sorting, associate each code with a number
nibrs_dict = dict(zip(nibrs, range(0,len(nibrs))))

# create a new column for the number
df['nibrs_num'] = df['nibrs_code'].apply(lambda x: nibrs_dict[x])

# group by the report number and pick the lowest NIBRS number if there are multiple events for the report number
df_filtered = df.sort('nibrs_num', ascending=True).groupby('report_number', as_index=False).first()

# filtered down from ~320,000 to ~250,000
df_filtered.info()

# this is the list of all crimes with suitable NIBRS codes from which we would like to match the addresses back to
df_filtered.to_csv('summit_crime_filtered_ungeocoded.csv',index_col=False)

###Check to make sure that we are filtering correctly

print 'The number of unique addresses to geocode is ' + str(len(set(df_filtered.address)))

Are we taking the correct NIBRS code for the report number?

df_filtered[df_filtered.report_number == df_filtered.report_number[100]]

df[df.report_number==df_filtered.report_number[100]]

Yes!

## Get the unique set of addresses we would like to geocode

# get the list of properties that are missing either a lat or a long

df_geo = df_filtered[(df_filtered['lat'].isnull()) | (df_filtered['long'].isnull())]

# get the crimes with unique addresses

df_geo = df_geo.sort('report_number', ascending=True).groupby('address', as_index=False).first()

df_geo.info()

So we have 42367 unique addresses to geocode...

df_geo.head(5)

Of which a number of them are quite bad...

def zip_string(x):
    try:
        x = str(int(x))
    except ValueError:
        x = str('')
    return x
df_geo['zip'] = df_geo['zip'].apply(zip_string) 
df_geo.address = df_geo.address.apply(str.upper)

print set(df_geo['zip'])

print len(set(df_geo.address))

df_geo.to_csv('data/geocode_me.csv', columns=['report_number','address','city','zip','lat','long'])
