==> 2_evaluate_geocode.ipynb <==
## Evaluate the geocode output

Geocoding was done using ArcMap and returned 41984 matched, 331 tied, and 52 unmatched for a *reported* 99% accuracy. 

import pandas as pd

# geocoded_crime.csv is a subset of addresses, as there were many repeated addresses!

df = pd.read_csv('data/geocoded.csv',index_col=[0])
df_geocodeme = pd.read_csv('data/geocode_me.csv', index_col=[0])

print len(df), len(df_geocodeme)

Let's check how many were geocoded in Ohio...

print "Percent geocoded in Ohio:", len(df[df.region=='Ohio'])*100.0/len(df)

This is clearly not up to the reported 99% accuracy, but still pretty good.

# df_filtered has the crime events filtered by nibrs_code, so we are ignoring a sizeable percent of events that had multiple charges
df_filtered = pd.read_csv('data/summit_crime_filtered_ungeocoded.csv', index_col = [0])

print len(set(df_filtered.address)), len(df.address)

Let's merge the filtered crime data with the addresses and see what our performance is.

# left join of the crime events and the geocoded address, on the address
df_geocoded = pd.merge(df_filtered, df[['address','x','y']],how='left',on=['address'])

df_geocoded.head()

Now let's put the geocoded lat/long (x/y) into the lat/fields.

# for events that got geocoded replace lat/long with the geocoded x/y
def replace(x):
    if not x['lat'] or x['long']:
        x['lat'] = x['x']
        x['long'] = x['y']
    return x
df_geocoded = df_geocoded.apply(replace, axis=1)

df_geocoded = df_geocoded[(~df_geocoded['lat'].isnull()) & (~df_geocoded['long'].isnull())]

print len(df_geocoded), len(df_filtered)

print("So we've really only geocoded {0:.2f}% of crime events.".format(231234.0/249568*100)) 

df_geocoded[['report_number','address','city','datetime_occurred','district','lat','long','nibrs_code','zip','nibrs_num']]\
    .to_csv('data/akron_crime_filtered_geocoded.csv',index=False)
